Renamed bunch of components
