+++ src/comps/feed/items/ItemCarteraUpdate.vue
@@ -89,7 +89,7 @@
 import client from '@/commons/client.api'
 
 export default {
-  name: 'Edit',
+  name: 'ItemCarteraUpdate',
 
   props: {
     edit: {
